Overhaul VFS
Instead of having a vfs_node tree, we now have mount_points, which
connect files in "host" file systems and root files in "guest" file
systems. This has the following benefits:

 - Previously it was possible to "mount" a file as a child of vfs_node
    without an associated directory. This created a weird situation that
    even though /dev/foo exists we failed to lookup /dev. Now our
    "mount" works just like on other Unix-like systems.
 - Tree traversing code was simplified.

Because of this change, we can no longer "mount" device files. Instead,
we use the ordinary way of representing device files with (major, minor)
numbers.

Since initrd is read-only, we cannot create /dev if initrd is the root
file system, so we switch the root file system to tmpfs and move initrd
to /initrd.